refactor impl suggestions from
