+++ rust-runtime/aws-smithy-types/src/date_time/mod.rs
@@ -15,7 +15,11 @@ use std::time::Duration;
 use std::time::SystemTime;
 use std::time::UNIX_EPOCH;
 
+#[cfg(all(test, aws_sdk_unstable, feature = "deserialize"))]
+mod de;
 mod format;
+#[cfg(all(test, aws_sdk_unstable, feature = "serialize"))]
+mod ser;
 
 pub use self::format::DateTimeFormatError;
 pub use self::format::DateTimeParseError;
@@ -322,121 +326,6 @@ pub enum Format {
     /// Number of seconds since the Unix epoch formatted as a floating point.
     EpochSeconds,
 }
-
-#[cfg(all(test, aws_sdk_unstable, feature = "serialize"))]
-mod ser {
-    use super::*;
-    use serde::ser::SerializeTuple;
-
-    impl serde::Serialize for DateTime {
-        fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-        where
-            S: serde::Serializer,
-        {
-            if serializer.is_human_readable() {
-                match self.fmt(Format::DateTime) {
-                    Ok(val) => serializer.serialize_str(&val),
-                    Err(e) => Err(serde::ser::Error::custom(e)),
-                }
-            } else {
-                let mut tup_ser = serializer.serialize_tuple(2)?;
-                tup_ser.serialize_element(&self.seconds)?;
-                tup_ser.serialize_element(&self.subsecond_nanos)?;
-                tup_ser.end()
-            }
-        }
-    }
-}
-
-#[cfg(all(test, aws_sdk_unstable, feature = "deserialize"))]
-mod der {
-    use super::*;
-    use serde::de::Visitor;
-    use serde::Deserialize;
-    struct DateTimeVisitor;
-
-    enum VisitorState {
-        First,
-        NotFirst,
-    }
-
-    struct NonHumanReadableDateTimeVisitor {
-        state: VisitorState,
-        seconds: i64,
-        subsecond_nanos: u32,
-    }
-
-    impl<'de> Visitor<'de> for DateTimeVisitor {
-        type Value = DateTime;
-        fn expecting(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-            formatter.write_str("expected RFC-3339 Date Time")
-        }
-
-        fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
-        where
-            E: serde::de::Error,
-        {
-            match DateTime::from_str(v, Format::DateTime) {
-                Ok(e) => Ok(e),
-                Err(e) => Err(serde::de::Error::custom(e)),
-            }
-        }
-    }
-
-    impl<'de> Visitor<'de> for NonHumanReadableDateTimeVisitor {
-        type Value = Self;
-        fn expecting(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-            formatter.write_str("expected (i64, u32)")
-        }
-
-        fn visit_i64<E>(mut self, v: i64) -> Result<Self::Value, E>
-        where
-            E: serde::de::Error,
-        {
-            match self.state {
-                VisitorState::First => {
-                    self.seconds = v;
-                    self.state = VisitorState::NotFirst;
-                }
-                _ => return Err(serde::de::Error::custom("First value must be i64")),
-            };
-            Ok(self)
-        }
-        fn visit_u32<E>(mut self, v: u32) -> Result<Self::Value, E>
-        where
-            E: serde::de::Error,
-        {
-            match self.state {
-                VisitorState::NotFirst => self.subsecond_nanos = v,
-                _ => return Err(serde::de::Error::custom("First value must be u32")),
-            };
-            Ok(self)
-        }
-    }
-
-    impl<'de> Deserialize<'de> for DateTime {
-        fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-        where
-            D: serde::Deserializer<'de>,
-        {
-            if deserializer.is_human_readable() {
-                deserializer.deserialize_str(DateTimeVisitor)
-            } else {
-                let visitor = NonHumanReadableDateTimeVisitor {
-                    state: VisitorState::First,
-                    seconds: 0,
-                    subsecond_nanos: 0,
-                };
-                let visitor = deserializer.deserialize_tuple(2, visitor)?;
-                Ok(DateTime {
-                    seconds: visitor.seconds,
-                    subsecond_nanos: visitor.subsecond_nanos,
-                })
-            }
-        }
-    }
-}
-
 #[cfg(test)]
 mod test {
     use crate::date_time::Format;
